CPM-548: Migrate product id to uuid (Step1 - database)

+++ src/Akeneo/Pim/Enrichment/Bundle/Elasticsearch/Indexer/ProductIndexer.php
@@ -5,9 +5,11 @@ declare(strict_types=1);
 namespace Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Indexer;
 
 use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\GetElasticsearchProductProjectionInterface;
+use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Model\ElasticsearchProductProjection;
 use Akeneo\Pim\Enrichment\Component\Product\Storage\Indexer\ProductIndexerInterface;
 use Akeneo\Tool\Bundle\ElasticsearchBundle\Client;
 use Akeneo\Tool\Bundle\ElasticsearchBundle\Refresh;
+use Ramsey\Uuid\UuidInterface;
 
 /**
  * Indexer responsible for the indexation of product entities. This indexer DOES NOT index product model ancestors that can
@@ -66,15 +68,7 @@ class ProductIndexer implements ProductIndexerInterface
                 $productIdentifiersChunk
             );
 
-            $normalizedProductProjections = (
-                static function (iterable $projections): iterable {
-                    foreach ($projections as $identifier => $projection) {
-                        yield $identifier => $projection->toArray();
-                    }
-                }
-            )($elasticsearchProductProjections);
-
-            $this->productAndProductModelClient->bulkIndexes($normalizedProductProjections, 'id', $indexRefresh);
+            $this->productAndProductModelClient->bulkIndexes($elasticsearchProductProjections, 'id', $indexRefresh);
         }
     }
 
@@ -95,11 +89,29 @@ class ProductIndexer implements ProductIndexerInterface
      */
     public function removeFromProductIds(array $productIds): void
     {
+        if (0 === count($productIds)) {
+            return;
+        }
+
         $this->productAndProductModelClient->bulkDelete(array_map(
             function ($productId) {
                 return self::PRODUCT_IDENTIFIER_PREFIX . (string) $productId;
             },
             $productIds
+        ));
+    }
+
+    public function removeFromProductUuids(array $productUuids): void
+    {
+        if (0 === count($productUuids)) {
+            return;
+        }
+
+        $this->productAndProductModelClient->bulkDelete(array_map(
+            function ($productUuid) {
+                return self::PRODUCT_IDENTIFIER_PREFIX . $productUuid->toString();
+            },
+            $productUuids
         ));
     }
 }
